RESP2REDIS: Refine error reporting from server (#25)
The api of RESP writer (which attached to RDB2RESP) is rather simple as it expects a plain iovec. It is not sufficient when it comes to play the RESP against live server such as the case of respToRedisLoader. In case of an error from the server it is desired to report an informative message to the user of the problematic command and key. librdb does have mechanism to report on the problematic command even in case of pipeline of commands but it reports the entire command as a raw data with not distinction down to command and key.

Extend API to pass as well command and key to RESP writer on start of a new RESP command.
(Another motivation to this change is to be aligned with RE Syncer code that expects specific formatted errors)
Refine and optimize ignoreChecksum flag.