RDB2RESP: Fix functions registration (#32)
In case of multiple levels registration of callbacks, RDB2RESP need to take
care to enforce that functions will be handled only by data-level (To simplify
implementation we avoid using FUNCTION RESTORE command and Instead, 
always use FUNCTION LOAD command. It is expected that there won't be 
many functions in the RDB file in which requires optimization).