Integrate python json parser as a service for testing (#46)
One of the outputs of librdb is JSON. Until now there was a naive testing code 
that compared two json files, attempting to ignore JSON elements ordering. This 
commit replaces it with a python service, run in the background, read JSON 
filename from its STDIN, generates its signature by ignoring JSON elements order, 
and prints SHA256 signature to its STDOUT. test_common.c will fork this service. 
The reason it is running as a service is to avoid the overhead of starting a new 
process for each evaluation and the intensive use of it.