Add support to hash-field-expiration RDB_TYPE v2 (#48)
RDB serialization of hashes with expiration on fields (aka. HFE) was modified
to write also at the begining the minimum expiration time.

Following this change, the new value will be digested but won't be propagated to
the callback handlers since it is actually a redundant information since each field
already "reports" its expiration time to handlers (This value will might be
used in the future for optimization in the context of dumping directly
from RDB to FLUSH without parsing the object).