Fix recv() 120s Timeout on EAGAIN by Retrying Indefinitely (#69)
Syncing with redis server may fail on configured 120-second timeout on recv()
in RESTORE MODE with a big key, say >4GB. It takes Redis server a long time to
process the RESTORE command before ACK. The solution is to modify the logic to
continuously retry without a timeout. An appropriate log message will be printed
every 10 seconds, such as: "No reply from Redis server for 10 seconds."

If requested, we can expose conifguration to fine tune timeout duration and retries.